requirements.txt
praw
tweepy
configparser
logging
python-telegram-bot
